+++ test/tests/Color.js
@@ -2,8 +2,8 @@
  * Paper.js - The Swiss Army Knife of Vector Graphics Scripting.
  * http://paperjs.org/
  *
- * Copyright (c) 2011 - 2019, Juerg Lehni & Jonathan Puckey
- * http://scratchdisk.com/ & https://puckey.studio/
+ * Copyright (c) 2011 - 2020, Jürg Lehni & Jonathan Puckey
+ * http://juerglehni.com/ & https://puckey.studio/
  *
  * Distributed under the MIT license. See LICENSE file for details.
  *
@@ -83,6 +83,9 @@ test('Creating Colors', function() {
 
     equals(new Color('rgba( 255, 0, 0, 0.5 )'), new Color(1, 0, 0, 0.5),
             'Color from rgba() string 2nd test');
+
+    equals(new Color('rgb(100%, 50%, 0%)'), new Color(1, 0.5, 0),
+            'Color from rgb() percenst string');
 
     equals(new Color('hsl(180deg, 20%, 40%)'),
             new Color({ hue: 180, saturation: 0.2, lightness: 0.4 }),