+++ test/tests/Raster.js
@@ -2,8 +2,8 @@
  * Paper.js - The Swiss Army Knife of Vector Graphics Scripting.
  * http://paperjs.org/
  *
- * Copyright (c) 2011 - 2019, Juerg Lehni & Jonathan Puckey
- * http://scratchdisk.com/ & https://puckey.studio/
+ * Copyright (c) 2011 - 2020, Jürg Lehni & Jonathan Puckey
+ * http://juerglehni.com/ & https://puckey.studio/
  *
  * Distributed under the MIT license. See LICENSE file for details.
  *
@@ -180,20 +180,26 @@ test('Raster#getAverageColor(path) with compound path', function() {
             { tolerance: 1e-3 });
 });
 
-test('Raster#smoothing defaults to true', function() {
+test('Raster#smoothing defaults to \'low\'', function() {
     var raster = new Raster();
-    equals(raster.smoothing, true);
+    equals(raster.smoothing, 'low');
 });
 
 test('Raster#smoothing', function() {
-    var raster = new Raster({ smoothing: false });
-    equals(raster.smoothing, false);
+    var raster = new Raster({ smoothing: 'off' });
+    equals(raster.smoothing, 'off');
+
+    raster.smoothing = 'medium';
+    equals(raster.smoothing, 'medium');
+
+    raster.smoothing = false;
+    equals(raster.smoothing, 'off');
 
     raster.smoothing = true;
-    equals(raster.smoothing, true);
+    equals(raster.smoothing, 'low');
 });
 
-test('Raster#setSmoothing setting does not impact canvas context', function(assert) {
+test('Raster#smoothing setting does not impact canvas context', function(assert) {
     var done = assert.async();
     var raster = new Raster('data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAIAAAACCAIAAAD91JpzAAAAGXRFWHRTb2Z0d2FyZQBBZG9iZSBJbWFnZVJlYWR5ccllPAAAABlJREFUeNpi+s/AwPCfgYmR4f9/hv8AAQYAHiAFAS8Lwy8AAAAASUVORK5CYII=');
     var view = raster.view;